% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_renderBodyPlaceholders.R
\name{checkBodyOutput}
\alias{checkBodyOutput}
\title{(internal function) Canned code to render placeholders for the critical check summary and results}
\usage{
checkBodyOutput()
}
\description{
(internal function) Canned code to render placeholders for the critical check summary and results
}
