% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createListOfFiles.R
\name{createListOfFiles}
\alias{createListOfFiles}
\title{Create a list of the R files in a specified folder}
\usage{
createListOfFiles(.folderUrl)
}
\arguments{
\item{.folderUrl}{Text url string to the folder containing the Registry Check datafiles}
}
\value{
A list of the names of the R files in the specified folder
}
\description{
Create a list of the R files in a specified folder
}
