% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_extractPassFail.R
\name{extractPassFail}
\alias{extractPassFail}
\title{(internal function) Extracts the critical check pass/fail results for a specific registrydqchecks dataset output}
\usage{
extractPassFail(.resultsList)
}
\arguments{
\item{.resultsList}{A list of the results in the specific format from registrydqchecks}
}
\value{
A list of the results of the critical check pass/fail results for the specific registrydqchecks dataset output
}
\description{
(internal function) Extracts the critical check pass/fail results for a specific registrydqchecks dataset output
}
