% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_extractPassFail.R
\name{extractPassFailUpper}
\alias{extractPassFailUpper}
\title{(internal function) Extract pass/fail results from all the datasets for this datapull}
\usage{
extractPassFailUpper(.dsList)
}
\arguments{
\item{.dsList}{A list containing the registrydqchecks output for all datasets for this datapull date}
}
\value{
A list of the pass/fail results for all critical checks for all datasets
}
\description{
(internal function) Extract pass/fail results from all the datasets for this datapull
}
