% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_extractRunnerSummary.R
\name{extractRunnerSummary}
\alias{extractRunnerSummary}
\title{Return runner summary information to print to the screen}
\usage{
extractRunnerSummary(.selectedPullInfo)
}
\arguments{
\item{.selectedPullInfo}{The name of the pull dataset and timestamp of the pull}
}
\value{
A summary of the person who ran the checks
}
\description{
Return runner summary information to print to the screen
}
