% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_getAllRdsFromPath.R
\name{getAllRdsFromPath}
\alias{getAllRdsFromPath}
\title{(internal function) Get a list of the names of all the R data files in a specified folder}
\usage{
getAllRdsFromPath(.folderUrl)
}
\arguments{
\item{.folderUrl}{A character url to the folder with the R datasets of interest}
}
\value{
A list containing the names of all the R datasets in the folder
}
\description{
(internal function) Get a list of the names of all the R data files in a specified folder
}
