% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_renderBodyPlaceholders.R
\name{renderBodyPlaceholders}
\alias{renderBodyPlaceholders}
\title{(internal function) Function to render a placeholder in the shiny app ui for the check results}
\usage{
renderBodyPlaceholders(.output)
}
\arguments{
\item{.output}{The shinyapp output object}
}
\description{
(internal function) Function to render a placeholder in the shiny app ui for the check results
}
