% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_renderBodyResults.R
\name{renderBodyResults}
\alias{renderBodyResults}
\title{(internal function) Render the check results in the body of the UI}
\usage{
renderBodyResults(.output, .summary, .checkList)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.summary}{The summary object to be rendered to the ui}

\item{.checkList}{The list of check results to be rendered to the ui}
}
\description{
(internal function) Render the check results in the body of the UI
}
