% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52_renderCCResultPlaceholder.R
\name{renderCCResultPlaceholder}
\alias{renderCCResultPlaceholder}
\title{(internal function) Render a placeholder for the specific critical check results on the shiny app ui}
\usage{
renderCCResultPlaceholder(.checkNum, .checkTitle)
}
\arguments{
\item{.checkNum}{The number for the specific critical check}

\item{.checkTitle}{The title for the specific critical check to be displayed}
}
\description{
(internal function) Render a placeholder for the specific critical check results on the shiny app ui
}
