% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_renderCCSummaryPlaceholder.R
\name{renderCCSummaryPlaceholder}
\alias{renderCCSummaryPlaceholder}
\title{(internal function) Render a placeholder for the critical check summary results on the shiny app ui}
\usage{
renderCCSummaryPlaceholder()
}
\description{
(internal function) Render a placeholder for the critical check summary results on the shiny app ui
}
