% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_renderCritCheck1.R
\name{renderCritCheck1}
\alias{renderCritCheck1}
\title{(internal function) Render the critical check 8 results to the ui}
\usage{
renderCritCheck1(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 8 results to the ui
}
