% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_renderCritCheck2.R
\name{renderCritCheck2}
\alias{renderCritCheck2}
\title{(internal function) Render the critical check 2 results to the ui}
\usage{
renderCritCheck2(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 2 results to the ui
}
