% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/84_renderCritCheck3.R
\name{renderCritCheck3}
\alias{renderCritCheck3}
\title{(internal function) Render the critical check 3 results to the ui}
\usage{
renderCritCheck3(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 3 results to the ui
}
