% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/85_renderCritCheck4.R
\name{renderCritCheck4}
\alias{renderCritCheck4}
\title{(internal function) Render the critical check 4 results to the ui}
\usage{
renderCritCheck4(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 4 results to the ui
}
