% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/86_renderCritCheck5.R
\name{renderCritCheck5}
\alias{renderCritCheck5}
\title{(internal function) Render the critical check 5 results to the ui}
\usage{
renderCritCheck5(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 5 results to the ui
}
