% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/87_renderCritCheck6.R
\name{renderCritCheck6}
\alias{renderCritCheck6}
\title{(internal function) Render the critical check 6 results to the ui}
\usage{
renderCritCheck6(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 6 results to the ui
}
