% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88_renderCritCheck7.R
\name{renderCritCheck7}
\alias{renderCritCheck7}
\title{(internal function) Render the critical check 7 results to the ui}
\usage{
renderCritCheck7(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 7 results to the ui
}
