% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/89_renderCritCheck8.R
\name{renderCritCheck8}
\alias{renderCritCheck8}
\title{(internal function) Render the critical check 8 results to the ui}
\usage{
renderCritCheck8(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the critical check to render}
}
\description{
(internal function) Render the critical check 8 results to the ui
}
