% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53_renderNCResultPlaceholder.R
\name{renderNCResultPlaceholder}
\alias{renderNCResultPlaceholder}
\title{(internal function) Render a placeholder for the noncritical check results on the shiny app ui}
\usage{
renderNCResultPlaceholder()
}
\description{
(internal function) Render a placeholder for the noncritical check results on the shiny app ui
}
