% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_renderNonCriticalCheckOutput.R
\name{renderNonCriticalCheckOuput}
\alias{renderNonCriticalCheckOuput}
\title{(internal function) Render the noncritical check results to the ui}
\usage{
renderNonCriticalCheckOuput(.output, .resultsToRender)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.resultsToRender}{The results of the noncritical checks to render}
}
\description{
(internal function) Render the noncritical check results to the ui
}
