% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/81_renderSummaryResults.R
\name{renderSummaryResults}
\alias{renderSummaryResults}
\title{(internal function) Render the summary results to the ui}
\usage{
renderSummaryResults(.output, .summary)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.summary}{The summary object to be rendered}
}
\description{
(internal function) Render the summary results to the ui
}
