% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApplication.R
\name{runApplication}
\alias{runApplication}
\title{Run the CorEvitas "Registry Data Quality Checks Report"}
\usage{
runApplication(.outputUrl)
}
\arguments{
\item{outputUrl}{A text url string to the location of the registry check output datasets}
}
\description{
Run the CorEvitas "Registry Data Quality Checks Report"
}
