% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_renderBodyResults.R
\name{renderBodyResults}
\alias{renderBodyResults}
\title{(internal function) Render the check results in the body of the UI}
\usage{
renderBodyResults(.output, .summary, .criticalCheckList, .nonCriticalCheckList)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.summary}{The summary object to be rendered to the ui}

\item{.criticalCheckList}{List of the critical checks to render}

\item{.nonCriticalCheckList}{List of the noncritical checks to render}
}
\description{
(internal function) Render the check results in the body of the UI
}
