% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_renderBodyResults.R
\name{renderCheckDetailResults}
\alias{renderCheckDetailResults}
\title{(internal function) Render the results of the critical checks to the ui}
\usage{
renderCheckDetailResults(.output, .criticalCheckList, .nonCriticalCheckList)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.checkList}{A list of the results of the critical checks}
}
\description{
(internal function) Render the results of the critical checks to the ui
}
