% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/81_renderSummaryResults.R
\name{renderSummaryResults}
\alias{renderSummaryResults}
\title{(internal function) Render the summary results to the ui}
\usage{
renderSummaryResults(.output, .ccSummary, .ncSummary)
}
\arguments{
\item{.output}{The shiny app server output object}

\item{.ccSummary}{The critical check summary object}

\item{.ncSummary}{The noncritical check summary object}
}
\description{
(internal function) Render the summary results to the ui
}
